Implement NoDeath feature and refine RemoteChest functionality
Adds the NoDeath feature to intercept and cancel death animations and respawn screen packets. Enhances RemoteChest with a deletion command and internal cleanup. Additionally, updates InventoryModule offsets to correctly identify the utility slot and applies various floating-point precision fixes across the renderer.

+++ Hytale/Renderer/FontRenderer/Font.cpp
@@ -113,5 +113,5 @@ float Font::getWidth(std::string text) {
 	return width;
 }
 float Font::getHeight() {
-	return m_size;
+	return (float)m_size;
 }
